requirements.txt
asgiref==3.8.1
awscli==1.41.3
boto==2.49.0
boto3==1.39.3
botocore==1.39.3
Brotli==1.1.0
build==1.2.2.post1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
crispy-bootstrap5==2025.4
dj-database-url==2.3.0
Django==5.2
django-anymail==13.0
django-autocomplete-light==3.12.1
django-crispy-forms==2.4
django-filter==25.1
django-ses-gateway==0.1.1
docutils==0.19
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
lxml==5.4.0
openpyxl==3.1.5
packaging==25.0
pip-tools==7.4.1
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.13.0
setuptools==80.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
whitenoise==6.9.0
